+++ src/view/provider/recording/helpers.ts
@@ -43,3 +43,16 @@ export const getBytesFromStorage = async (): Promise<number> => {
     }
     return 0;
 };
+
+export const backgroundStopRecordingBytes = async () => {
+    const tabs = await chrome.tabs.query({ active: true, currentWindow: true });
+    if (tabs.length > 0) {
+        const activeTabId = tabs[0].id;
+        if (activeTabId) {
+            chrome.runtime.sendMessage({
+                command: "stopRecordingBytesTransferred",
+                tabId: activeTabId,
+            });
+        }
+    }
+};
